Remove admin/superadmin login/sigup, and approval
