--- src/project/ui/AdminSignupAndLoginMenu.java
@@ -1,90 +1,0 @@
-package project.ui;
-
-import project.service.UserService;
-
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.WindowConstants;
-import java.awt.Image;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
-
-public class AdminSignupAndLoginMenu implements ActionListener {
-    private final JFrame launchPageFrame;
-    private final JFrame coverPageFrame;
-    private final UserService userService;
-    private final JFrame frame = new JFrame("Admin");
-    private final JLabel label = new JLabel();
-    private final JButton loginButton = new JButton("Log in");
-    private final JButton signupButton = new JButton("Sign up");
-    private final JButton exitButton = new JButton("EXIT");
-    private String windowEventSource = "";
-
-    public AdminSignupAndLoginMenu(JFrame launchPageFrame, JFrame coverPageFrame, final UserService userService) {
-        this.launchPageFrame = launchPageFrame;
-        this.coverPageFrame = coverPageFrame;
-        this.userService = userService;
-
-        // Import logo for the Customer frame
-        ImageIcon icon = new ImageIcon("beach2.png");
-
-        // Import background for the Customer frame
-        ImageIcon background = new ImageIcon("beach3.jpg");
-        Image backgroundImage = background.getImage().getScaledInstance(350, 300, Image.SCALE_DEFAULT);
-        JLabel backgroundLabel = new JLabel(new ImageIcon(backgroundImage));
-        backgroundLabel.setBounds(0, 0, 350, 300);
-
-        loginButton.setBounds(75, 50, 200, 40);
-        loginButton.setFocusable(false);
-        loginButton.addActionListener(this);
-
-        signupButton.setBounds(75, 100, 200, 40);
-        signupButton.setFocusable(false);
-        signupButton.addActionListener(this);
-
-        exitButton.setBounds(75, 150, 200, 40);
-        exitButton.setFocusable(false);
-        exitButton.addActionListener(this);
-
-        frame.add(label);
-        frame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-        frame.setSize(350, 300);
-        frame.setIconImage(icon.getImage());
-        frame.add(loginButton);
-        frame.add(signupButton);
-        frame.add(exitButton);
-        frame.add(backgroundLabel);
-        frame.setLayout(null);
-        frame.setResizable(false);
-        frame.setLocationRelativeTo(null);
-        frame.setVisible(true);
-        frame.addWindowListener(new WindowAdapter() {
-            @Override
-            public void windowClosed(WindowEvent e) {
-                if (!"loginButton".equals(windowEventSource) && !"signupButton".equals(windowEventSource)) {
-                    coverPageFrame.setVisible(true);
-                }
-            }
-        });
-    }
-
-    @Override
-    public void actionPerformed(ActionEvent e) {
-        if (e.getSource() == loginButton) {
-            windowEventSource = "loginButton";
-            frame.dispose();
-            new AdminLogin(launchPageFrame, userService);
-        } else if (e.getSource() == signupButton) {
-            windowEventSource = "signupButton";
-            frame.dispose();
-            new AdminSignup(launchPageFrame, userService);
-        } else {
-            frame.dispose();
-            coverPageFrame.setVisible(true);
-        }
-    }
-}
